menu/serializers/transaction.py
from rest_framework import serializers

from menu.models import Transaction


class TransactionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Transaction
        fields = '__all__'
